+++ src/sll/include/sll/scheduler.h
@@ -178,14 +178,14 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_reset_barrier(sll_barrier_index
 
 
 /**
- * \flags func
+ * \flags check_output func
  * \name sll_start_thread
  * \group scheduler
  * \desc Docs!
  * \arg sll_thread_index_t t
  * \ret sll_bool_t
  */
-__SLL_EXTERNAL sll_bool_t sll_start_thread(sll_thread_index_t t);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_start_thread(sll_thread_index_t t);